+++ src/LiveChartsCore/AssemblyInfo.cs
@@ -28,7 +28,7 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.Avalonia")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.WinForms")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.WinUI")]
-[assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.UWP")]
+[assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.Uno")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.WPF")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.XamarinForms")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.Eto")]
@@ -46,7 +46,7 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.Avalonia,       PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.WinForms,       PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.WinUI,          PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
-[assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.UWP,            PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
+[assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.Uno,            PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.WPF,            PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.XamarinForms,   PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
 [assembly: InternalsVisibleTo("LiveChartsCore.SkiaSharpView.Eto,            PublicKey=0024000004800000940000000602000000240000525341310004000001000100b5845b22b111de8f9504daf136f975d790af5361b9612a3dea48fc3480087efa128adbfe0267ee4f894d8985e8b9c55f6734896e192fa647e142e1aadf281131d2329b345e456bbb6ff5b3ef4bf8da574022d0e9618a68c9241a61350bb505b6ae6a49f6c3d45be7aa76f5646574d40870c22d648d57765e040cac6e8b21c3d3")]
