+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/DefaultLegend.xaml
@@ -1,9 +1,9 @@
 ﻿<UserControl
-    x:Class="LiveChartsCore.SkiaSharpView.UWP.DefaultLegend"
+    x:Class="LiveChartsCore.SkiaSharpView.Uno.DefaultLegend"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:core="using:LiveChartsCore"
-    xmlns:local="using:LiveChartsCore.SkiaSharpView.UWP"
+    xmlns:local="using:LiveChartsCore.SkiaSharpView.Uno"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
