+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/GeoMap.xaml
@@ -1,14 +1,16 @@
 ﻿<UserControl
-    x:Class="LiveChartsCore.SkiaSharpView.UWP.MotionCanvas"
+    x:Class="LiveChartsCore.SkiaSharpView.Uno.GeoMap"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LiveChartsCore.SkiaSharpView.UWP"
+    xmlns:local="using:LiveChartsCore.SkiaSharpView.Uno"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:skia="using:SkiaSharp.Views.UWP"
     mc:Ignorable="d"
     d:DesignHeight="300"
     d:DesignWidth="400">
 
-    <skia:SKXamlCanvas x:Name="canvas"></skia:SKXamlCanvas>
+    <Grid>
+        <local:MotionCanvas x:Name="canvas" >
+        </local:MotionCanvas>
+    </Grid>
 </UserControl>
