+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/MotionCanvas.xaml
@@ -1,16 +1,14 @@
 ﻿<UserControl
-    x:Class="LiveChartsCore.SkiaSharpView.UWP.GeoMap"
+    x:Class="LiveChartsCore.SkiaSharpView.Uno.MotionCanvas"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LiveChartsCore.SkiaSharpView.UWP"
+    xmlns:local="using:LiveChartsCore.SkiaSharpView.Uno"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:skia="using:SkiaSharp.Views.UWP"
     mc:Ignorable="d"
     d:DesignHeight="300"
     d:DesignWidth="400">
 
-    <Grid>
-        <local:MotionCanvas x:Name="canvas" >
-        </local:MotionCanvas>
-    </Grid>
+    <skia:SKXamlCanvas x:Name="canvas"></skia:SKXamlCanvas>
 </UserControl>
