+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/PieChart.xaml
@@ -1,8 +1,8 @@
 ﻿<UserControl
-    x:Class="LiveChartsCore.SkiaSharpView.UWP.CartesianChart"
+    x:Class="LiveChartsCore.SkiaSharpView.Uno.PieChart"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LiveChartsCore.SkiaSharpView.UWP"
+    xmlns:local="using:LiveChartsCore.SkiaSharpView.Uno"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
@@ -26,7 +26,7 @@
             <RowDefinition Height="*"></RowDefinition>
             <RowDefinition Height="Auto"></RowDefinition>
         </Grid.RowDefinitions>
-        <local:MotionCanvas Grid.Row="1" Grid.Column="1" x:Name="motionCanvas" />
-        <local:DefaultLegend Grid.Row="1" Grid.Column="2" x:Name="legend" VerticalAlignment="Center" HorizontalAlignment="Center" />
+        <local:MotionCanvas Grid.Row="1" Grid.Column="1" x:Name="motionCanvas"/>
+        <local:DefaultLegend Grid.Row="1" Grid.Column="2" x:Name="legend" VerticalAlignment="Center" HorizontalAlignment="Center"/>
     </Grid>
 </UserControl>
