+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/PolarChart.xaml
@@ -1,8 +1,8 @@
 ﻿<UserControl
-    x:Class="LiveChartsCore.SkiaSharpView.UWP.PolarChart"
+    x:Class="LiveChartsCore.SkiaSharpView.Uno.PolarChart"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LiveChartsCore.SkiaSharpView.UWP"
+    xmlns:local="using:LiveChartsCore.SkiaSharpView.Uno"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
