+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/Properties/LiveChartsCore.SkiaSharpView.UWP.rd.xml
