+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/TooltipBindingContext.cs
@@ -20,49 +20,23 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Windows.UI.Text;
+using System;
 using Windows.UI.Xaml.Media;
 
-namespace LiveChartsCore.SkiaSharpView.UWP
+namespace LiveChartsCore.SkiaSharpView.Uno;
+
+/// <summary>
+/// Defines the tooltip binding context class.
+/// </summary>
+public class TooltipBindingContext
 {
     /// <summary>
-    /// Defines the binding series class.
+    /// Gets the background.
     /// </summary>
-    public class BindingSeries
-    {
-        /// <summary>
-        /// Gets the series.
-        /// </summary>
-        public ISeries Series { get; set; }
+    public Brush Background { get; set; } = null!;
 
-        /// <summary>
-        /// Gets the font family.
-        /// </summary>
-        public FontFamily FontFamily { get; set; }
-
-        /// <summary>
-        /// Gets the foreground.
-        /// </summary>
-        public Brush Foreground { get; set; }
-
-        /// <summary>
-        /// Gets the font size.
-        /// </summary>
-        public double FontSize { get; set; }
-
-        /// <summary>
-        /// Gets the font weight.
-        /// </summary>
-        public FontWeight FontWeight { get; set; }
-
-        /// <summary>
-        /// Gets the font style.
-        /// </summary>
-        public FontStyle FontStyle { get; set; }
-
-        /// <summary>
-        /// Gets the font stretch.
-        /// </summary>
-        public FontStretch FontStretch { get; set; }
-    }
+    /// <summary>
+    /// Gets the points.
+    /// </summary>
+    public BindingPoint[] Points { get; set; } = Array.Empty<BindingPoint>();
 }
